refactor code to fs service
