+++ src/api/common.ts
@@ -1,5 +1,4 @@
 import { ipcRenderer } from 'electron';
-import fs from 'promise-fs';
 
 export const sendNotification = (content: string) => {
     ipcRenderer.send('send-notification', content);
@@ -10,12 +9,3 @@ export const showOnTray = (content: string) => {
 export const reloadWindow = () => {
     ipcRenderer.send('reload-window');
 };
-
-export async function doesFolderExists(dirPath: string) {
-    return await fs
-        .stat(dirPath)
-        .then((stats) => {
-            return stats.isDirectory();
-        })
-        .catch(() => false);
-}
